requirements.txt
Django==1.8
django-filter==0.11.0
djangorestframework==3.2.3
django-cors-headers==1.1.0
gunicorn==19.3.0
MySQL-python==1.2.5
requests==2.9.1
django-crispy-forms=1.6.1